memorry corruption/unsorted chunks
Minimal is in principle as Minimal2 without the reading of data. The error occurs by extracting values from a symmetric matrix into a array created with "new". By providing only the already generated matrix there is no error but by creating the matrix when reading in data the error occurs when assigning the values to the array.